refactor: 移动方法位置
