+++ src/ui/index.ts
@@ -27,5 +27,4 @@ export { default as OptionRadioGroup } from '@/ui/OptionRadioGroup.vue'
 export { type RadioItem } from '@/ui/radioItem'
 export { default as OptionWidgetLayout } from '@/ui/OptionWidgetLayout.vue'
 export { default as RadioGroupItem } from '@/ui/RadioGroupItem.vue'
-export { createComponentWithProps } from '@/ui/vue-component-factory'
 export { default as MiniToast } from '@/core/toast/MiniToast.vue'
